Fix some merge mistakes.

+++ src/estimator/dcm_attitude_estimator.cpp
@@ -139,11 +139,11 @@ AttitudeEstimate DCMAttitudeEstimator::makeEstimate(const SensorMeasurements& me
     // TODO(yoos): Due to singularities, this is really only safe for
     // visualization. We can use this for position control but should not
     // deviate more than perhaps 30 degrees away from horizontal.
-    .roll = atan2(dcm(1,2), dcm(2,2)),
-    .pitch = asin(-dcm(0,2)),
+    .roll = atan2f(dcm(1,2), dcm(2,2)),
+    .pitch = asinf(-dcm(0,2)),
     //.roll = -atan2(dcm(1,2), dcm(2,2)) * dcm(0,0) + atan2(dcm(0,2), dcm(2,2)) * dcm(1,0),
     //.pitch = atan2(dcm(0,2), dcm(2,2)) * dcm(1,1) - atan2(dcm(1,2), dcm(2,2)) * dcm(0,1),
-    .yaw = atan2(dcm(0,1), dcm(0,0)),
+    .yaw = atan2f(dcm(0,1), dcm(0,0)),
 
     // Velocities are set later if a gyro is available.
     .rollVel = 0.0f,
